# contracts/gateway-bff.openapi.yaml
openapi: 3.1.0
info:
  title: foculist gateway-bff API
  version: 0.1.0
components:
  parameters:
    tenantHeader:
      in: header
      name: X-Tenant-ID
      required: true
      schema:
        type: string
      description: Tenant identifier for request isolation
...
paths:
...
  /api/sync/push:
    post:
      summary: Gateway route to sync-service push endpoint
      parameters:
        - $ref: '#/components/parameters/tenantHeader'
      responses:
        "200":
          description: Proxied sync push response
  /api/sync/pull:
    post:
      summary: Gateway route to sync-service pull endpoint
      parameters:
        - $ref: '#/components/parameters/tenantHeader'
      responses:
        "200":
          description: Proxied sync pull response
...